Reorganize the tests in IncrementalCompilationTests.swift
# Conflicts:
#	Tests/SwiftDriverTests/IncrementalCompilationTests.swift
